requirements.txt
altgraph==0.10.2
amqp==2.3.2
ansible==2.6.2
appnope==0.1.0
asn1crypto==0.24.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
bdist-mpkg==0.5.0
billiard==3.5.0.4
bonjour-py==0.3
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
decorator==4.3.0
Django==1.11.2
django-daterange-filter==1.3.0
django-filter==1.1.0
djangorestframework==3.8.2
enum34==1.1.6
futures==3.2.0
httpie==0.9.9
idna==2.7
ipaddress==1.0.22
ipython==5.7.0
ipython-genutils==0.2.0
Jinja2==2.10
kombu==4.2.1
macholib==1.5.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==1.3.1
modulegraph==0.10.4
MySQL-python==1.2.5
nose==1.3.7
numpy==1.8.0rc1
paramiko==2.4.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py2app==0.7.3
pyasn1==0.4.4
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
PyNaCl==1.2.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
PyYAML==3.13
redis==2.10.6
requests==2.19.1
scandir==1.7
scipy==0.13.0b1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
tornado==5.0.2
traitlets==4.3.2
urllib3==1.23
vboxapi==1.0
vine==1.1.4
wcwidth==0.1.7
xattr==0.6.4
zope.interface==4.1.1
